Use the current `op` to determine which `etuple` function to use
The first argument passed to `etuple` determines which of the registered
functions is called. The first argument is generally an
`ExpressionTuple`, the etuplized version of the operator. However, when
using subclasses of `ExpressionTuple` to customize evaluation, we want
`etuple` to depend on the current operator and not its etuplized
version.